Add comments Some bugs are fixed Add new NamedQuery for User (User.findByUserLoginAndPassword)

+++ src/main/java/com/opensource/model/UsersBooksPK.java
@@ -9,16 +9,24 @@ import java.io.Serializable;
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
 
+
 /**
+ * Project <b> elib</b>. 
  *
- * @author Alexandr
+ * This Embeddable describes a composite primary key for UsersBooks class
+ *
+ * @version $version$
+ * @author Oleksandr Lukichov
+ *
+ * @since 18 זמגע. 2015 נ.
+ *
  */
 @Embeddable
 public class UsersBooksPK implements Serializable{
 
     @Column(name = "user_id")
-    private int userId;
+    private Integer userId;
 
     @Column(name = "book_id")
-    private Long bookId;
+    private Integer bookId;
 }
